Refactor engine module
