+++ configs/task.yaml
@@ -1,6 +1,6 @@
 # @package _group_
 create_gif: 0
-create_plots: 1
+create_plots: 0
 generate_samples: 1
 n_generate_samples: 10
 n_plot_samples: 5
